mtools (4.0.33-1+really4.0.32-1) unstable; urgency=medium

  * Revert to version 4.0.32-1 due to regressions on ARM* systems affecting the
    the Debian Installer. Thanks to Cyril Brulebois (kibi) for the report and
    patience. (Closes: #991403)

mtools (4.0.33-1) unstable; urgency=medium

  * New upstream release.

mtools (4.0.32-1) unstable; urgency=medium

  * New upstream release.

mtools (4.0.31-1) unstable; urgency=medium

  * New upstream release.

mtools (4.0.30-1) unstable; urgency=medium

  * New upstream release.

mtools (4.0.29-1) unstable; urgency=medium

  * New upstream release.

mtools (4.0.27-1) unstable; urgency=medium

  * New upstream release.
  * Apply wrap-and-sort -sa.

+++ copyfile.c
@@ -24,11 +24,11 @@
  * Copy the data from source to target
  */
 
-int copyfile(Stream_t *Source, Stream_t *Target)
+ssize_t copyfile(Stream_t *Source, Stream_t *Target)
 {
 	char buffer[8*16384];
 	mt_off_t pos;
-	int ret;
+	ssize_t ret;
 	ssize_t retw;
 /*	size_t len;*/
 	mt_size_t mt_len;
@@ -63,8 +63,8 @@ int copyfile(Stream_t *Source, Stream_t *Target)
 				perror("write in copy");
 			else
 				fprintf(stderr,
-					"Short write %lu instead of %d\n",
-					(unsigned long) retw, ret);
+					"Short write %zd instead of %zd\n",
+					retw, ret);
 			if(errno == ENOSPC)
 				got_signal = 1;
 			return ret;
